+++ dataset_creator/creator.py
@@ -77,6 +77,8 @@ class Creator(object):
         else:  # TNT
             return tnt.TntDatasetBlock(self.data, self.codon_positions,
                                        self.partitioning,
+                                       degenerate=self.degenerate,
+                                       aminoacids=self.aminoacids,
                                        outgroup=self.outgroup).dataset_block()
 
     def create_dataset_footer(self):